Second pass of cleaning up passes API.
