--- codepropertygraph/src/main/scala/io/shiftleft/diffgraph/AppliedDiffGraph.scala
@@ -1,23 +1,0 @@
-package io.shiftleft.diffgraph
-
-import java.util
-
-import io.shiftleft.codepropertygraph.generated.nodes.NewNode
-import org.apache.tinkerpop.gremlin.structure.Vertex
-import java.lang.{Long => JLong}
-
-/**
-  * Diff Graph that has been applied to a source graph. This is a wrapper around
-  * diff graph, which additionally provides a map from nodes to graph ids.
-  * */
-case class AppliedDiffGraph(diffGraph: DiffGraph,
-                            private val nodeToTinkerNode: util.HashMap[IdentityHashWrapper[NewNode], Vertex]) {
-
-  /**
-    * Obtain the id this node has in the applied graph
-    * */
-  def nodeToGraphId(node: NewNode): JLong = {
-    val wrappedNode = IdentityHashWrapper(node)
-    nodeToTinkerNode.get(wrappedNode).id.asInstanceOf[JLong]
-  }
-}
