--- codepropertygraph/src/main/scala/io/shiftleft/diffgraph/IdentityHashWrapper.scala
@@ -1,12 +1,0 @@
-package io.shiftleft.diffgraph
-
-private[diffgraph] case class IdentityHashWrapper[T <: AnyRef](value: T) {
-  override def hashCode(): Int = {
-    System.identityHashCode(value)
-  }
-
-  override def equals(other: Any): Boolean =
-    other != null &&
-      other.isInstanceOf[IdentityHashWrapper[T]] &&
-      (this.value eq other.asInstanceOf[IdentityHashWrapper[T]].value)
-}
