+++ semanticcpg/src/main/scala/io/shiftleft/passes/languagespecific/fuzzyc/TypeDeclStubCreator.scala
@@ -2,8 +2,7 @@ package io.shiftleft.passes.languagespecific.fuzzyc
 import gremlin.scala.ScalaGraph
 import io.shiftleft.codepropertygraph.Cpg
 import io.shiftleft.codepropertygraph.generated.{NodeTypes, nodes}
-import io.shiftleft.diffgraph.DiffGraph
-import io.shiftleft.passes.CpgPass
+import io.shiftleft.passes.{CpgPass, DiffGraph}
 
 /**
   * This pass has no other pass as prerequisite.
