+++ semanticcpg/src/main/scala/io/shiftleft/passes/propagateedges/PropagateEdgePass.scala
@@ -3,8 +3,7 @@ package io.shiftleft.passes.propagateedges
 import gremlin.scala._
 import io.shiftleft.semanticsloader.Semantics
 import io.shiftleft.codepropertygraph.generated._
-import io.shiftleft.diffgraph.DiffGraph
-import io.shiftleft.passes.CpgPass
+import io.shiftleft.passes.{CpgPass, DiffGraph}
 import org.apache.logging.log4j.{LogManager, Logger}
 import org.apache.tinkerpop.gremlin.structure.Direction
