+++ semanticcpg/src/main/scala/io/shiftleft/passes/receiveredges/ReceiverEdgePass.scala
@@ -3,8 +3,7 @@ package io.shiftleft.passes.receiveredges
 import gremlin.scala._
 import io.shiftleft.codepropertygraph.generated.nodes.StoredNode
 import io.shiftleft.codepropertygraph.generated.{EdgeTypes, NodeKeys, NodeTypes}
-import io.shiftleft.diffgraph.DiffGraph
-import io.shiftleft.passes.CpgPass
+import io.shiftleft.passes.{CpgPass, DiffGraph}
 import org.apache.logging.log4j.{LogManager, Logger}
 import org.apache.tinkerpop.gremlin.structure.Direction
